Move CacheBackend port from persistence to indexing ports
- Define CacheBackend Protocol in _indexing/ports/cache_backend.py.
- SqliteCacheBackend remains the persistence-side implementation.
- Update cache_manager.py to import from the inner port location.
- Keep _persistence/cache_backend.py as a temporary re-export.

make format-check, make lint, make type-check pass.
Full test suite: 1432 passed, 17 skipped, 4 xpassed.
