+++ clang_index_mcp/_indexing/ports/__init__.py
@@ -1,0 +1,5 @@
+"""Ports (interfaces) for the indexing use-case layer."""
+
+from .cache_backend import CacheBackend
+
+__all__ = ["CacheBackend"]
