+++ clang_index_mcp/_persistence/cache_manager.py
@@ -9,7 +9,7 @@ from collections import defaultdict
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Union
 
-from .._persistence.cache_backend import CacheBackend
+from .._indexing.ports.cache_backend import CacheBackend
 from .._persistence.cache_validation_context import CacheValidationContext
 from .._core.error_tracking import ErrorTracker, RecoveryManager
 from .._persistence.project_identity import ProjectIdentity
